+++ CHANGELOG.md
@@ -9,6 +9,8 @@ use patch releases for compatibility fixes instead.
 
 ## Unreleased
 
+## [v1.14.0](https://github.com/allenai/beaker-py/releases/tag/v1.14.0) - 2023-01-17
+
 ### Added
 
 - Added `prefix` parameter to `Beaker.dataset.ls()`.